third_party/blink/web_tests/platform/generic/fast/events/ime-composition-events-001-expected.txt

This tests that calling input-method functions sends Composition Events and Text Events introduced in DOM Level 3. To test manually, enable an IME, input CJK characters, and see this page doesn't show 'FAIL' lines.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS event.type is "compositionstart"
PASS event.data is ""
PASS event.type is "compositionupdate"
PASS event.data is "1"
PASS event.type is "compositionupdate"
PASS event.data is "2"
PASS event.type is "compositionupdate"
PASS event.data is "3"
PASS event.type is "compositionupdate"
PASS event.data is "4"
PASS event.type is "textInput"
PASS event.data is "4"
PASS event.type is "compositionend"
PASS event.data is "4"
PASS event.type is "compositionstart"
PASS event.data is ""
PASS event.type is "compositionupdate"
PASS event.data is "5"
PASS event.type is "compositionupdate"
PASS event.data is "6"
PASS event.type is "compositionupdate"
PASS event.data is "7"
PASS event.type is "compositionupdate"
PASS event.data is ""
PASS event.type is "textInput"
PASS event.data is ""
PASS event.type is "compositionend"
PASS event.data is ""
PASS event.type is "textInput"
PASS event.data is "8"
PASS event.type is "compositionstart"
PASS event.data is ""
PASS event.type is "compositionupdate"
PASS event.data is "9"
PASS event.type is "compositionend"
PASS event.data is "9"
PASS event.type is "compositionstart"
PASS event.data is "have"
PASS event.type is "compositionupdate"
PASS event.data is "lost"
PASS test.value is "I lost a pen"
PASS event.type is "compositionupdate"
PASS event.data is "made"
PASS event.type is "textInput"
PASS event.data is "made"
PASS event.type is "compositionend"
PASS event.data is "made"
PASS test.value is "I made a pen"
PASS successfullyParsed is true

TEST COMPLETE
